refactor: complete package.json exports and add new direct export files
Updated package.json exports to use direct file exports instead of index.ts:
- packages/chat-ai-tools: export ./tools only
- packages/chat-core: export ./types only
- packages/chat-services: export individual service files
- packages/chat-trpc: removed index export

Added new direct export files:
- packages/chat-ai-tools/src/tools.ts
- packages/chat-core/src/types.ts

This completes the migration to direct exports for better tree shaking and clearer server/client code separation.
