+++ packages/chat-ai-tools/package.json
@@ -4,8 +4,8 @@
   "private": true,
   "type": "module",
   "exports": {
-    ".": {
-      "default": "./src/index.ts"
+    "./tools": {
+      "default": "./src/tools.ts"
     }
   },
   "scripts": {
