+++ packages/chat-ai-tools/src/tools.ts
