+++ packages/chat-core/package.json
@@ -4,9 +4,9 @@
   "private": true,
   "type": "module",
   "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./src/index.ts"
+    "./types": {
+      "types": "./dist/types.d.ts",
+      "default": "./src/types.ts"
     }
   },
   "license": "MIT",
