+++ packages/chat-trpc/package.json
@@ -4,9 +4,6 @@
   "private": true,
   "type": "module",
   "exports": {
-    ".": {
-      "default": "./src/index.ts"
-    },
     "./client": {
       "default": "./src/client.ts"
     },
