--- packages/chat-trpc/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./client";
-export * from "./server";
-export * from "./react";
